[V2] Admin: Update content
