+++ sites/www/src/routes/(admin)/admin/content/[id]/schema.ts
@@ -7,6 +7,7 @@ export const schema = z.object({
     slug: z.string().min(1, 'Slug is required'),
     description: z.string().min(1, 'Description is required'),
     tags: z.array(z.number()).min(1, 'At least one tag is required'),
+    status: z.enum(['draft', 'published']).default('draft'),
     metadata: z.union([
         z.object({
             videoId: z.string().regex(/^[a-zA-Z0-9_-]+$/)
